# server/public/docs/api-docs.json
{
    "info": {
        "title": "iamdocuman API",
        "version": "1.0.0",
        "description": "Demonstrating how to describe a RESTful API with Swagger"
    },
    "swagger": "2.0",
    "paths": {
        "/api/v1/documents/?limit=4&offset=2": {
            "get": {
                "description": "Returns {limit} documents from the the {offset}",
                "tags": [
                    "Get documents"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get documents from the database",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Document"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/documents": {
            "post": {
                "description": "Creates a document",
                "tags": [
                    "Create Document"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "description": "User object",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/NewDocument"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "object,",
                            "items": {
                                "$ref": "#/definitions/Document"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/documents/:id": {
            "get": {
                "description": "Returns {limit} documents from the the {offset}",
                "tags": [
                    "Get single document"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get documents from the database",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Document"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update  a document",
                "tags": [
                    "Update a Document"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "description": "User object",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/NewDocument"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "object,",
                            "items": {
                                "$ref": "#/definitions/Document"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/documents/1": {
            "delete": {
                "description": "Deletes the document with the id of 1",
                "tags": [
                    "Delete document"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Document"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "description": "Returns all users",
                "tags": [
                    "Get users"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a user",
                "tags": [
                    "Create User"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "description": "User object",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "object,",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/?limit=4&offset=2": {
            "get": {
                "description": "Returns {limit} users from the the {offset}",
                "tags": [
                    "Get users"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/1": {
            "get": {
                "description": "Returns the user with the id of 1",
                "tags": [
                    "Get user"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Creates a user",
                "tags": [
                    "Update User"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "description": "User object",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "object,",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the user with the id of 1",
                "tags": [
                    "Delete user"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/login": {
            "post": {
                "description": "Logs in a user",
                "tags": [
                    "Login User"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "description": "User object",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/NewLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "type": "object,",
                            "items": {
                                "$ref": "#/definitions/Login"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/1/documents": {
            "get": {
                "description": "Returns the documents belonging to the user of id 1",
                "tags": [
                    "Get Documents of A User"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user's documents",
                        "schema": {
                            "type": "array"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "NewDocument": {
            "type": "object",
            "required": [
                "title",
                "content"
            ],
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "text"
                }
            }
        },
        "Document": {
            "allOf": [
                {
                    "$ref": "#/definitions/NewDocument"
                },
                {
                    "required": [
                        "id"
                    ]
                },
                {
                    "properties": null,
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            ]
        },
        "NewUser": {
            "type": "object",
            "required": [
                "firstname",
                "lastname",
                "username",
                "email",
                "password"
            ],
            "properties": {
                "firstname": {
                    "type": "string",
                    "example": "Han"
                },
                "lastname": {
                    "type": "string",
                    "example": "Solo"
                },
                "username": {
                    "type": "string",
                    "example": "g-pirate"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "example": "millenium-falcon"
                },
                "email": {
                    "type": "string",
                    "example": "hansolo@documan.api"
                }
            }
        },
        "User": {
            "allOf": [
                {
                    "$ref": "#/definitions/NewUser"
                },
                {
                    "required": [
                        "id"
                    ]
                },
                {
                    "properties": null,
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            ]
        },
        "NewLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                }
            }
        },
        "Login": {
            "allOf": [
                {
                    "$ref": "#/definitions/NewLogin"
                }
            ]
        }
    },
    "responses": {},
    "parameters": {},
    "securityDefinitions": {},
    "tags": []
}